Refactored native_instance_wraper
